scrapping/pyproject.toml
[project]
name = "scrapping"
version = "0.1.0"
description = "web scrapping tool(samad)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "fastapi>=0.95.0",
    "httpx>=0.24.0",
    "openai>=0.27.0",
    "playwright>=1.36.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.24.0",
    "uvicorn>=0.22.0",
   
]
